atualizaca

+++ src/main/java/com/senac/pedrorogerpedroroger/repository/PedidoRepository.java
@@ -8,5 +8,5 @@ import java.util.List;
 
 @Repository
 public interface PedidoRepository extends JpaRepository<Pedido, Long> {
-    List<Pedido> findByUsuarioUsuarioId(Long usuarioId);
+    List<Pedido> findByUsuarioUsuarioId(int usuarioId);
 }
